refactor: split cli argument helpers
